Refactor new_trace(), finish_trace(), live_span(), setenv() and END block. Replace TRACE_RE constant with $TRACE_RE.
Grouped declaration of globals $trace_id, @live_span and @span_ids together.
They seem related. Someone should add a comment.

The constant TRACE_RE was a constant for no good reason. And it required an
explicit `$_ =~ TRACE_RE` which was irksome.

Also, this part of regex `0x[0-9a-zA-Z]{1,16}` may be hexadecimal. If so, `a-fA-F`
would be better.